+++ grove-web/src/components/Tasks/TaskDetail/TaskDetail.tsx
@@ -63,8 +63,10 @@ export function TaskDetail({
       <TaskTabs activeTab={activeTab} onTabChange={setActiveTab} />
 
       {/* Tab Content */}
-      <div className="flex-1 overflow-y-auto p-4">
-        {renderTabContent()}
+      <div className="flex-1 min-h-0 p-4 flex flex-col">
+        <div className="flex-1 min-h-0 overflow-y-auto">
+          {renderTabContent()}
+        </div>
       </div>
 
       {/* Actions */}